requirements.txt
flask==3.1.0
google-generativeai==0.8.4
twilio==9.4.0
python-dotenv==1.0.1
gunicorn==23.0.0
psycopg2-binary==2.9.9
